Fix minor int conversion errors

+++ src/custom_lib_functions/lib_functions.h
@@ -612,7 +612,7 @@ namespace lib_functions
 
 		osu_object_v<T> obj_v_c;
 
-		for (size_t i = offset; i < obj_v->size(); i += n) {
+        for (int i = offset; i < obj_v->size(); i += n) {
 			obj_v_c.push_back(obj_v->get_index(i));
 		}
 
@@ -632,7 +632,7 @@ namespace lib_functions
 		// off = 1
 		// n = 3
 
-		for (size_t i = offset; i < obj_v->size(); i ++) {
+        for (int i = offset; i < obj_v->size(); i ++) {
 			if ((i - offset) % n != 0) { // Only push back those not in the nth sequence
 				obj_v_c.push_back(obj_v->get_index(i));
 			}
